Revert "[Tizen] Add codes for Dali Windows Backend"
This reverts commit 4e33556657b8d5fd5ba75cf1cbba45f473b97742.

+++ dali-toolkit/internal/layouting/linear-layout-impl.cpp
@@ -280,7 +280,7 @@ void LinearLayout::LayoutHorizontal( LayoutLength left, LayoutLength top, Layout
   auto height = bottom - top;
 
   // Space available for child
-  auto childSpace = height - (int)padding.top - (int)padding.bottom;
+  auto childSpace = height - padding.top - padding.bottom;
 
   auto count = GetChildCount();
 
@@ -305,11 +305,11 @@ void LinearLayout::LayoutHorizontal( LayoutLength left, LayoutLength top, Layout
       auto childHeight = childLayout->GetMeasuredHeight();
       auto childMargin = childLayout->GetMargin();
 
-      childTop = LayoutLength(padding.top) + (int)((childSpace - childHeight) / 2) + (int)childMargin.top - (int)childMargin.bottom;
+      childTop = LayoutLength(padding.top) + ((childSpace - childHeight) / 2) + childMargin.top - childMargin.bottom;
 
       childLeft += childMargin.start;
       childLayout->Layout( childLeft, childTop, childLeft + childWidth, childTop + childHeight );
-      childLeft += childWidth + (int)childMargin.end + mCellPadding.width;
+      childLeft += childWidth + childMargin.end + mCellPadding.width;
     }
   }
 }
@@ -440,7 +440,7 @@ void LinearLayout::LayoutVertical( LayoutLength left, LayoutLength top, LayoutLe
   auto width = right - left;
 
   // Space available for child
-  auto childSpace = width - (int)padding.start - (int)padding.end;
+  auto childSpace = width - padding.start - padding.end;
   auto count = GetChildCount();
 
   for( unsigned int childIndex = 0; childIndex < count; childIndex++)
@@ -456,7 +456,7 @@ void LinearLayout::LayoutVertical( LayoutLength left, LayoutLength top, LayoutLe
       childLeft = LayoutLength( padding.start ) + ( childSpace - childWidth ) / 2 + childMargin.start - childMargin.end;
 
       childLayout->Layout( childLeft, childTop, childLeft + childWidth, childTop + childHeight );
-      childTop += childHeight + (int)childMargin.bottom + mCellPadding.height;
+      childTop += childHeight + childMargin.bottom + mCellPadding.height;
     }
   }
 }
